+++ CHANGELOG.md
@@ -20,6 +20,8 @@
      * `stubbles\webapp\Response::notAcceptable()`
      * `stubbles\webapp\Response::httpVersionNotSupported()`
    * both `stubbles\webapp\auth\User` and `stubbles\webapp\auth\Roles` are not available via injection any more, use `stubbles\webapp\Request::identity()` instead
+   * post interceptors of auth protected resources are now always called even when actual request is not authorized
+
 
 ### Other changes